+++ src/Haskollider/Bus.hs
@@ -12,16 +12,16 @@ isSettable :: Bus -> Bool
 isSettable bus = (rate bus) /= Audio
 
 set :: Bus -> Double -> Message
-set bus value = Message "/c_set" [int32 (index bus), double value]
+set bus value = Message "/c_set" [int32 (index bus), float value]
 
 setn :: Bus -> [Double] -> Message
-setn bus values = Message "/c_setn" $ [int32 (index bus), int32 (length values)] ++ (foldr (\x acc -> (double x) : acc) [] values)
+setn bus values = Message "/c_setn" $ [int32 (index bus), int32 (length values)] ++ (foldr (\x acc -> (float x) : acc) [] values)
 
 setAt :: Bus -> Int -> [Double] -> Message
-setAt bus offset values = Message "/c_set" $ [int32 (index bus + offset), int32 (length values)] ++ (foldr (\x acc -> (double x) : acc) [] values)
+setAt bus offset values = Message "/c_set" $ [int32 (index bus + offset), int32 (length values)] ++ (foldr (\x acc -> (float x) : acc) [] values)
 
 fill :: Bus -> Double -> Int -> Message
-fill bus value nChannels = Message "/c_fill" [int32 (index bus), int32 nChannels, double value]
+fill bus value nChannels = Message "/c_fill" [int32 (index bus), int32 nChannels, float value]
 
 setAll :: Bus -> Double -> Message
 setAll bus value =  fill bus value (numChannels bus)
